requirements-dev.txt
asyncpg-stubs
black
coverage[toml]
maubot
mypy
pytest
pytest-asyncio
pytest-aiohttp
pytest-cov
respx
ruff
types-pytz
